requirements.txt
pandas
scikit-learn
torch
fastapi
uvicorn
pydantic
pyyaml
llama-cpp-python
sentence-transformers
chromadb
requests
tqdm